+++ R/bCond.tree.conversion.R
@@ -127,9 +127,13 @@ treeCKT2matrixInd <- function(estimatedTree, newDataXJ = NULL)
 #'
 matrixInd2matrixCKT <- function(matrixInd, newDataXI)
 {
-  if (nrow(newDataXI) != nrow(matrixInd)){
-    stop("`newDataXI` and `matrixInd` should have the same number of rows, ",
-         "as they come from the same sample.")
+  if (NROW(newDataXI) != NROW(matrixInd)){
+    stop(errorCondition(
+      message = paste0("`newDataXI` and `matrixInd` should have the same number",
+                       "of rows, as they come from the same sample. ",
+                       "Here they are respectively: ",
+                       NROW(newDataXI), ", ", NROW(matrixInd)),
+      class = "DifferentLengthsError") )
   }
 
   # We count the number of boxes in the partition